+++ breeze-dsl/src/main/kotlin/com/windea/breezeframework/dsl/graph/mermaid/MermaidStateDiagram.kt
@@ -1,4 +1,4 @@
-@file:Suppress("NOTHING_TO_INLINE")
+@file:Suppress("NOTHING_TO_INLINE", "unused")
 
 package com.windea.breezeframework.dsl.graph.mermaid
 
@@ -6,25 +6,22 @@ import com.windea.breezeframework.core.annotations.api.*
 import com.windea.breezeframework.core.extensions.*
 import com.windea.breezeframework.dsl.*
 import com.windea.breezeframework.dsl.graph.mermaid.MermaidConfig.indent
-import com.windea.breezeframework.dsl.graph.mermaid.MermaidConfig.quote
-import org.intellij.lang.annotations.*
-
-//TODO
 
 //NOTE unstable raw api
 
 //REGION top annotations and interfaces
 
-/**PlantUml状态图的Dsl。*/
+/**Mermaid状态图的Dsl。*/
 @ReferenceApi("[Mermaid State Diagram](https://mermaidjs.github.io/#/stateDiagram)")
 @DslMarker
 private annotation class MermaidStateDiagramDsl
 
-/**PlantUml状态图。*/
+/**Mermaid状态图。*/
 @MermaidStateDiagramDsl
 class MermaidStateDiagram @PublishedApi internal constructor() : Mermaid(), MermaidStateDiagramDslEntry, CanIndentContent {
 	override val states: MutableSet<MermaidStateDiagramState> = mutableSetOf()
 	override val links: MutableList<MermaidStateDiagramTransition> = mutableListOf()
+	override val notes: MutableList<MermaidStateDiagramNote> = mutableListOf()
 	
 	override var indentContent: Boolean = true
 	override var splitContent: Boolean = true
@@ -35,18 +32,21 @@ class MermaidStateDiagram @PublishedApi internal constructor() : Mermaid(), Merm
 	}
 }
 
+
 //REGION dsl interfaces
 
-/**PlantUml状态图Dsl的入口。*/
+/**Mermaid状态图Dsl的入口。*/
 @MermaidStateDiagramDsl
 interface MermaidStateDiagramDslEntry : MermaidDslEntry, CanSplitContent, WithTransition<MermaidStateDiagramState, MermaidStateDiagramTransition> {
 	val states: MutableSet<MermaidStateDiagramState>
 	val links: MutableList<MermaidStateDiagramTransition>
+	val notes: MutableList<MermaidStateDiagramNote>
 	
 	fun _toContentString(): String {
 		return arrayOf(
 			states.joinToStringOrEmpty("\n"),
-			links.joinToStringOrEmpty("\n")
+			links.joinToStringOrEmpty("\n"),
+			notes.joinToStringOrEmpty("\n")
 		).filterNotEmpty().joinToStringOrEmpty(_splitWrap)
 	}
 	
@@ -54,7 +54,7 @@ interface MermaidStateDiagramDslEntry : MermaidDslEntry, CanSplitContent, WithTr
 	override fun String.fromTo(other: String) = transition(this, other)
 }
 
-/**PlantUml状态图Dsl的元素。*/
+/**Mermaid状态图Dsl的元素。*/
 @MermaidStateDiagramDsl
 interface MermaidStateDiagramDslElement : MermaidDslElement
 
@@ -66,58 +66,33 @@ sealed class MermaidStateDiagramState(
 	val name: String,
 	val text: String? = null
 ) : MermaidStateDiagramDslElement, WithName {
-	var alias: String? = null
+	override val _name: String get() = name
 	
-	override val _name: String get() = alias ?: name
+	override fun equals(other: Any?) = equalsBySelect(this, other) { arrayOf(name) }
 	
-	override fun equals(other: Any?) = equalsBySelect(this, other) { arrayOf(alias ?: name) }
-	
-	override fun hashCode() = hashCodeBySelect(this) { arrayOf(alias ?: name) }
+	override fun hashCode() = hashCodeBySelect(this) { arrayOf(name) }
 }
 
-/**
- * PlantUml状态图简单状态。
- *
- * 语法示例：
- * ```
- * state A: text can\n wrap line.
- *
- * state A<<Tag>> #333: state with tag and color.
- *
- * state "Long name" as A: state with long name.
- * ```
- */
+/**Mermaid状态图简单状态。*/
 @MermaidStateDiagramDsl
 class MermaidStateDiagramSimpleState @PublishedApi internal constructor(
 	name: String,
 	text: String? = null
 ) : MermaidStateDiagramState(name, text) {
+	var type: Type? = null
+	
 	override fun toString(): String {
-		val textSnippet = text?.replaceWithEscapedWrap().orEmpty()
-		return if(alias == null) {
-			"state $name: $textSnippet"
-		} else {
-			val nameSnippet = name.replaceWithEscapedWrap().wrapQuote(quote)
-			"state $nameSnippet as $alias: $textSnippet"
-		}
+		val textSnippet = text?.let { ": $it" }.orEmpty()
+		val typeSnippet = type?.let { " <<${it.text}>>" }.orEmpty()
+		return "$name$textSnippet$typeSnippet"
+	}
+	
+	enum class Type(internal val text: String) {
+		Fork("fork"), Join("join")
 	}
 }
 
-/**
- * Mermaid状态图复合状态。
- *
- * 语法示例：
- * ```
- * state A {
- *     A1 -> A2
- * }
- *
- * state A { ... }
- * state A: text can\n wrap line.
- *
- * state A { ... }
- * state "Long name" as A: state with long name.
- */
+/**Mermaid状态图复合状态。*/
 @MermaidStateDiagramDsl
 class MermaidStateDiagramCompositedState @PublishedApi internal constructor(
 	name: String,
@@ -125,60 +100,30 @@ class MermaidStateDiagramCompositedState @PublishedApi internal constructor(
 ) : MermaidStateDiagramState(name, text), MermaidStateDiagramDslEntry, CanIndentContent {
 	override val states: MutableSet<MermaidStateDiagramState> = mutableSetOf()
 	override val links: MutableList<MermaidStateDiagramTransition> = mutableListOf()
+	override val notes: MutableList<MermaidStateDiagramNote> = mutableListOf()
 	
 	override var indentContent: Boolean = true
 	override var splitContent: Boolean = true
 	
 	override fun toString(): String {
-		val contentSnippet = _toContentString()
-			.let { if(indentContent) it.prependIndent(indent) else it }
-		val nameSnippet = alias ?: name
-		val extraSnippet = alias?.let { "\nstate ${name.replaceWithEscapedWrap().wrapQuote(quote)} as $alias" }
-			.orEmpty()
-		val extraSnippetWithText = when {
-			text == null -> extraSnippet
-			extraSnippet.isEmpty() -> "\nstate $name: ${text.replaceWithEscapedWrap()}"
-			else -> "$extraSnippet: ${text.replaceWithEscapedWrap()}"
-		}
-		return "state $nameSnippet {\n$contentSnippet\n}$extraSnippetWithText"
+		val contentSnippet = _toContentString()._applyIndent(indent)
+		return "state $name {\n$contentSnippet\n}"
 	}
 }
 
-/**
- * Mermaid状态图并发状态。
- *
- * 语法示例：
- * ```
- * state A {
- *     A1 -> A2
- *     ---
- *     A2 -> A1
- * }
- */
+/**Mermaid状态图并发状态。*/
 @MermaidStateDiagramDsl
 class MermaidStateDiagramConcurrentState @PublishedApi internal constructor(
 	name: String,
 	text: String? = null
 ) : MermaidStateDiagramState(name, text), CanIndentContent {
-	val states: MutableSet<MermaidStateDiagramSimpleState> = mutableSetOf()
 	val sections: MutableList<MermaidStateDiagramConcurrentSection> = mutableListOf()
 	
 	override var indentContent: Boolean = true
 	
 	override fun toString(): String {
-		val contentSnippet = arrayOf(
-			states.joinToStringOrEmpty("\n"),
-			sections.joinToStringOrEmpty("\n---\n")
-		).filterNotEmpty().joinToStringOrEmpty("\n\n")
-		val indentedContentSnippet = if(indentContent) contentSnippet.prependIndent(indent) else contentSnippet
-		val nameSnippet = alias ?: name
-		val extraSnippet = alias?.let { "\nstate ${name.replaceWithHtmlWrap().wrapQuote(quote)} as $alias" }.orEmpty()
-		val extraSnippetWithText = when {
-			text == null -> extraSnippet
-			extraSnippet.isEmpty() -> "\nstate $name: ${text.replaceWithHtmlWrap()}"
-			else -> "$extraSnippet: ${text.replaceWithHtmlWrap()}"
-		}
-		return "state $nameSnippet {\n$indentedContentSnippet\n}$extraSnippetWithText"
+		val contentSnippet = sections.joinToStringOrEmpty("\n--\n")._applyIndent(indent)
+		return "state $name {\n$contentSnippet\n}"
 	}
 }
 
@@ -187,51 +132,58 @@ class MermaidStateDiagramConcurrentState @PublishedApi internal constructor(
 class MermaidStateDiagramConcurrentSection : MermaidStateDiagramDslEntry {
 	override val states: MutableSet<MermaidStateDiagramState> = mutableSetOf()
 	override val links: MutableList<MermaidStateDiagramTransition> = mutableListOf()
+	override val notes: MutableList<MermaidStateDiagramNote> = mutableListOf()
 	
 	override var splitContent: Boolean = true
 	
 	override fun toString() = _toContentString()
 }
 
-/**
- * PlantUml状态图转换。
- *
- * 语法示例：
- * * `A -> B`
- * * `A -> B: text can\n wrap line.`
- * * `A --> B: arrow can be longer.`
- * * `A -[#333,dotted]right-> B: arrow with color, shape and direction.`
- */
+/**Mermaid状态图转换。*/
 @MermaidStateDiagramDsl
 class MermaidStateDiagramTransition @PublishedApi internal constructor(
 	val sourceStateName: String,
 	val targetStateName: String,
-	@Language("Creole")
-	var text: String? = null //NOTE can wrap by "\n"
+	var text: String? = null
 ) : MermaidStateDiagramDslElement, WithNode<MermaidStateDiagramState> {
-	var arrowColor: String? = null
-	var arrowShape: ArrowShape? = null
-	
-	var arrowDirection: MermaidArrowDirection? = null
-	
-	var arrowLength: Int = 1
-	
 	override val _fromNodeName get() = sourceStateName
 	override val _toNodeName get() = targetStateName
 	
 	override fun toString(): String {
-		val textSnippet = text?.let { ": ${it.replaceWithEscapedWrap()}" }.orEmpty()
-		val arrowParamsSnippet = arrayOf(arrowColor?.addPrefix("#"), arrowShape?.text).filterNotNull()
-			.joinToStringOrEmpty(",", "[", "]")
-		val arrowDirectionSnippet = arrowDirection?.text.orEmpty()
-		val arrowLengthSnippet = "-" * (arrowLength - 1)
-		val arrowSnippet = "-$arrowParamsSnippet$arrowDirectionSnippet$arrowLengthSnippet>"
-		return "$sourceStateName $arrowSnippet $targetStateName$textSnippet"
+		val textSnippet = text?.let { ": $it" }.orEmpty()
+		return "$sourceStateName --> $targetStateName$textSnippet"
+	}
+}
+
+class MermaidStateDiagramNote @PublishedApi internal constructor(
+	val location: Location,
+	val text: String //NOTE can be multiline
+) : MermaidStateDiagramDslElement, CanWrapContent, CanIndentContent {
+	override var wrapContent: Boolean = false
+	override var indentContent: Boolean = true
+	
+	override fun toString(): String {
+		if("\n" in text || "\r" in text) wrapContent = true //wrap if necessary
+		
+		return when {
+			wrapContent && indentContent -> "note $location\n${text._applyIndent(indent)}\nend note"
+			wrapContent -> "note $location\n$text\nend note"
+			else -> "note $location: $text"
+		}
 	}
 	
-	enum class ArrowShape(internal val text: String)
+	class Location @PublishedApi internal constructor(
+		val position: Position,
+		val stateName: String
+	) {
+		override fun toString(): String {
+			return "${position.text} $stateName"
+		}
+	}
 	
-	enum class MermaidArrowDirection(internal val text: String)
+	enum class Position(internal val text: String) {
+		LeftOf("left of"), RightOf("right of")
+	}
 }
 
 //REGION build extensions
@@ -245,13 +197,11 @@ inline fun MermaidStateDiagramDslEntry.state(name: String, text: String? = null)
 
 @InlineDsl
 @MermaidStateDiagramDsl
-inline val MermaidStateDiagramDslEntry.initState
-	get() = "[*]"
+inline fun MermaidStateDiagramDslEntry.initState() = "[*]"
 
 @InlineDsl
 @MermaidStateDiagramDsl
-inline val MermaidStateDiagramDslEntry.finishState
-	get() = "[*]"
+inline fun MermaidStateDiagramDslEntry.finishState() = "[*]"
 
 @MermaidStateDiagramDsl
 inline fun MermaidStateDiagramDslEntry.compositedState(name: String, text: String? = null,
@@ -271,42 +221,31 @@ inline fun MermaidStateDiagramDslEntry.transition(sourceStateName: String, targe
 @MermaidStateDiagramDsl
 inline fun MermaidStateDiagramDslEntry.transition(sourceState: MermaidStateDiagramSimpleState,
 	targetState: MermaidStateDiagramSimpleState, text: String? = null) =
-	MermaidStateDiagramTransition(sourceState.alias ?: sourceState.name, targetState.alias ?: targetState.name, text)
+	MermaidStateDiagramTransition(sourceState.name, targetState.name, text)
 		.also { links += it }
 
 @MermaidStateDiagramDsl
-inline fun MermaidStateDiagramDslEntry.transition(
-	sourceStateName: String,
-	arrowColor: String?,
-	arrowStyle: MermaidStateDiagramTransition.ArrowShape?,
-	arrowDirection: MermaidStateDiagramTransition.MermaidArrowDirection?,
-	targetStateName: String,
-	text: String? = null
-) = MermaidStateDiagramTransition(sourceStateName, targetStateName, text)
-	.also { it.arrowColor = arrowColor;it.arrowShape = arrowStyle;it.arrowDirection = arrowDirection }
-	.also { links += it }
+inline fun MermaidStateDiagramDslEntry.note(location: MermaidStateDiagramNote.Location, text: String) =
+	MermaidStateDiagramNote(location, text).also { notes += it }
 
-
+@InlineDsl
 @MermaidStateDiagramDsl
-inline infix fun MermaidStateDiagramState.alias(alias: String) =
-	this.also { it.alias = alias }
+inline fun MermaidStateDiagramDslEntry.leftOf(stateName: String) =
+	MermaidStateDiagramNote.Location(MermaidStateDiagramNote.Position.LeftOf, stateName)
 
+@InlineDsl
 @MermaidStateDiagramDsl
-inline fun MermaidStateDiagramConcurrentState.state(name: String, text: String? = null) =
-	MermaidStateDiagramSimpleState(name, text).also { states += it }
+inline fun MermaidStateDiagramDslEntry.rightOf(stateName: String) =
+	MermaidStateDiagramNote.Location(MermaidStateDiagramNote.Position.RightOf, stateName)
 
 @MermaidStateDiagramDsl
 inline fun MermaidStateDiagramConcurrentState.section(block: MermaidStateDiagramConcurrentSection.() -> Unit) =
 	MermaidStateDiagramConcurrentSection().also { it.block() }.also { sections += it }
 
 @MermaidStateDiagramDsl
+inline infix fun MermaidStateDiagramSimpleState.type(type: MermaidStateDiagramSimpleState.Type) =
+	this.also { it.type = type }
+
+@MermaidStateDiagramDsl
 inline infix fun MermaidStateDiagramTransition.text(text: String) =
 	this.also { it.text = text }
-
-@MermaidStateDiagramDsl
-inline infix fun MermaidStateDiagramTransition.arrowColor(arrowColor: String) =
-	this.also { it.arrowColor = arrowColor }
-
-@MermaidStateDiagramDsl
-inline infix fun MermaidStateDiagramTransition.arrowShape(arrowShape: MermaidStateDiagramTransition.ArrowShape) =
-	this.also { it.arrowShape = arrowShape }